Implement review recommendations
